ai-photo-platform/backend/requirements.txt
uvicorn==0.27.0
fastapi==0.109.0
python-multipart==0.0.6
sqlalchemy==2.0.25
asyncpg>=0.31.0
redis==5.0.1
celery==5.3.6
insightface==0.7.3
rembg>=2.0.68
opencv-python==4.9.0.80
numpy>=2.0.0
pillow>=10.0.0
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.1.2
aiofiles==23.2.1
python-dotenv==1.0.0
httpx==0.26.0
